fix(proxy) resolve DNS hostnames post-plugins
This is mainly to address the Lambda and OpenWhisk plugins not being
able to run if the specified ùpstream_url`'s hostname does not resolve,
hence resulting in an HTTP 500 error.

Ideally, this change should probably be reverted, since some future
plugins might need to know the resolved IP address or the upstream Host
header in their `access` phase.